Promote the complete runtime-proven implementation tree

+++ apps/launcher/datafilespage.cpp
@@ -145,7 +145,7 @@ namespace Launcher
 
         int getMaxNavMeshDbFileSizeMiB()
         {
-            return static_cast<int>(Settings::navigator().mMaxNavmeshdbFileSize / (1024 * 1024));
+            return Settings::navigator().mMaxNavmeshdbFileSize / (1024 * 1024);
         }
     }
 }
