+++ apps/openmw/mwgui/postprocessorhud.cpp
@@ -172,16 +172,16 @@ namespace MWGui
         if (selected == MyGUI::ITEM_NONE)
             return;
 
-        size_t index = direction == Direction::Up ? selected - 1 : selected + 1;
-        index = std::clamp<size_t>(index, 0, mActiveList->getItemCount() - 1);
+        int index = direction == Direction::Up ? static_cast<int>(selected) - 1 : selected + 1;
+        index = std::clamp<int>(index, 0, mActiveList->getItemCount() - 1);
 
-        if (index != selected)
+        if (static_cast<size_t>(index) != selected)
         {
             auto technique = getTechnique(*mActiveList, selected);
             if (technique->getDynamic() || technique->getInternal())
                 return;
 
-            if (processor->enableTechnique(std::move(technique), static_cast<int>(index) - mOffset)
+            if (processor->enableTechnique(std::move(technique), index - mOffset)
                 != MWRender::PostProcessor::Status_Error)
                 processor->saveChain();
         }
@@ -279,7 +279,7 @@ namespace MWGui
 
     void PostProcessorHud::notifyMouseWheel(MyGUI::Widget* /*sender*/, int rel)
     {
-        double offset = mConfigLayout->getViewOffset().top + rel * 0.3;
+        int offset = mConfigLayout->getViewOffset().top + rel * 0.3;
         if (offset > 0)
             mConfigLayout->setViewOffset(MyGUI::IntPoint(0, 0));
         else
