+++ apps/openmw/mwrender/postprocessor.hpp
@@ -224,7 +224,7 @@ namespace MWRender
 
         void updateLiveReload();
 
-        void cull(unsigned frameId, osgUtil::CullVisitor* cv);
+        void cull(size_t frameId, osgUtil::CullVisitor* cv);
 
         osg::ref_ptr<osg::Group> mRootNode;
         osg::ref_ptr<osg::Camera> mHUDCamera;
@@ -246,7 +246,7 @@ namespace MWRender
 
         size_t mDirtyFrameId = 0;
         size_t mLastFrameNumber = 0;
-        double mLastSimulationTime = 0.0;
+        float mLastSimulationTime = 0.f;
 
         bool mDirty = false;
         bool mReload = true;
