+++ apps/openmw/mwscript/globalscripts.hpp
@@ -76,7 +76,7 @@ namespace MWScript
         void addStartup();
         ///< Add startup script
 
-        size_t countSavedGameRecords() const;
+        int countSavedGameRecords() const;
 
         void write(ESM::ESMWriter& writer, Loading::Listener& progress) const;
