+++ apps/openmw/mwworld/globals.cpp
@@ -100,7 +100,7 @@ namespace MWWorld
         }
     }
 
-    size_t Globals::countSavedGameRecords() const
+    int Globals::countSavedGameRecords() const
     {
         return mVariables.size();
     }
