+++ apps/openmw/mwworld/worldmodel.hpp
@@ -97,7 +97,7 @@ namespace MWWorld
 
         std::vector<MWWorld::Ptr> getAll(const ESM::RefId& id);
 
-        size_t countSavedGameRecords() const;
+        int countSavedGameRecords() const;
 
         void write(ESM::ESMWriter& writer, Loading::Listener& progress) const;
